add enable/disable option and timeout config

+++ msg_defs.py
@@ -10,7 +10,7 @@ class ModelParameter(BaseModel):
     top_k: int | None = None
     top_p: float | None = None
     max_new_tokens: int | None = 300
-    max_time: float | None = None
+    max_time: int | None = None
 
 
 class ChatRequest(BaseModel):
